CSS tweak to add padding for survivor data below the map
When looking through hexes for survivors, clicking on an empty hex removes
the survivor data entirely, causing the map to move due to the lack of
entities below the map. Increasing the vertical size of the page means
there's empty space so the map stays in place when the survivor data
vanishes while scrolled down to see survivors.