Full Survivor List (#8)
* Fix off-by-one bug on total item counts

* Adding survivor positions to details list, hooked into map click

Split the callback for updating the Data field with survivor information
out into a standalone function so that it can be easily called
elsewhere. One step closer to an interactive list of survivors

* Adding touchscreen support, wrapping for smaller screens

I haven't quite gotten the survivor info to wrap properly on my phone,
each individual line wraps over (e.g. breaking "Name:" and the name onto
different lines) before it actually starts a new row.

* Removing accidental blank line

* Emergency parser fix since loose items/structures were added

* Parsing loose items and structures as survivors.

This may not be the cleanest way to store them, but it does mean that
all the existing tools for examining survivors just work, and especially
for structures it does make sense to list as a survivor-controlled
entity. Parsing is wrapped in try-catch to guard against future breaks.

* Adding 'all survivors' button to dump all survivor info on the page.

With the css tweaks to ensure survivors wrap properly, it's not
unreasonable to just dump the entire list. Helps when looking for
survivors with empty inventory or HP damage, which were the main reasons
I was still opening the Status post to check survivor details.